+++ src/main/java/com/enigma/repository/impl/FilmRepoImpl.java
@@ -1,12 +1,10 @@
 package com.enigma.repository.impl;
 
 import com.enigma.entity.Film;
-import com.enigma.entity.Rating;
 import com.enigma.repository.FilmRepo;
 
 import java.sql.*;
 import java.util.ArrayList;
-import java.util.List;
 
 public class FilmRepoImpl implements FilmRepo {
     Connection conn;
@@ -16,22 +14,22 @@ public class FilmRepoImpl implements FilmRepo {
     }
 
     @Override
-    public List<Film> getAll() {
-        List<Film> data = new ArrayList<>();
+    public void getAll() {
+        ArrayList<Film> data = new ArrayList<>();
         try{
             PreparedStatement pr = conn.prepareStatement("select tf.*, tr.code from t_film tf join t_rating tr on tf.rating_id = tr.id;");
             ResultSet result = pr.executeQuery();
             while (result.next()) {
                 data.add(new Film(result.getInt("id"),result.getString("title"), result.getInt("duration"), result.getString("show_date"),result.getInt("price"), result.getInt("rating_id"), result.getString("code")));
             }
+            getList(data);
         } catch (SQLException e) {
             System.out.println(e.getMessage());
         }
-        return data;
     }
 
     @Override
-    public void getById(Integer id) {
+    public Film getById(Integer id) {
         Film film = null;
         try {
             PreparedStatement pr = conn.prepareStatement("select tf.*, tr.code from t_film tf join t_rating tr on tf.rating_id = tr.id where tf.id =" + id + ";");
@@ -39,11 +37,12 @@ public class FilmRepoImpl implements FilmRepo {
             while (result.next()) {
                 film = new Film(result.getInt("id"),result.getString("title"), result.getInt("duration"), result.getString("show_date"),result.getInt("price"), result.getInt("rating_id"), result.getString("code"));
             }
+            System.out.println(film == null ? "Film not found" : "Found " + film);
             pr.close();
-            System.out.println("Found : " + film );
         } catch (SQLException e){
             System.out.println("Failed get data by Id " + e.getMessage());
         }
+        return film;
     }
 
     @Override
@@ -57,7 +56,8 @@ public class FilmRepoImpl implements FilmRepo {
             pr.setInt(5, film.getRatingId());
 
             pr.executeUpdate();
-            System.out.println("success add new data : " + film.getTitle());
+            System.out.println("success add new film : title= " + film.getTitle() + ", duration= " + film.getDuration() + ", Date= " + film.getShowDate() +", Price= " + film
+                    .getPrice());
             pr.close();
         } catch (SQLException e) {
             System.out.println("failed save data : "+e.getMessage());
@@ -67,37 +67,60 @@ public class FilmRepoImpl implements FilmRepo {
 
     @Override
     public void update(Film film) {
+        Film films;
+        films = getById(film.getId());
         try{
-            PreparedStatement pr = conn.prepareStatement("update t_film set title=?, duration=?, show_date=?, price=?, rating_id=? where id=?;");
-            pr.setString(1, film.getTitle());
-            pr.setInt(2, film.getDuration());
-            pr.setDate(3,Date.valueOf(film.getShowDate()));
-            pr.setInt(4,film.getPrice());
-            pr.setInt(5,film.getRatingId());
-            pr.setInt(6, film.getId());
+            if (films == null) {
+                System.out.println("Cannot update film");
+            } else {
+                PreparedStatement pr = conn.prepareStatement("update t_film set title=?, duration=?, show_date=?, price=?, rating_id=? where id=?;");
+                pr.setString(1, film.getTitle());
+                pr.setInt(2, film.getDuration());
+                pr.setDate(3,Date.valueOf(film.getShowDate()));
+                pr.setInt(4,film.getPrice());
+                pr.setInt(5,film.getRatingId());
+                pr.setInt(6, film.getId());
 
-            int updated = pr.executeUpdate();
-            if (updated > 0) System.out.println("success update data");
-            else System.out.println("no data updated");
-            pr.close();
-
+                int updated = pr.executeUpdate();
+                if (updated > 0) System.out.println("success update film " + film.getTitle());
+                else System.out.println("no data film updated");
+                pr.close();
+            }
         } catch (SQLException e) {
-            System.out.println("Failed update : " +e.getMessage());
+            System.out.println("Failed update film : " +e.getMessage());
         }
     }
 
     @Override
     public void delete(Integer id) {
+        Film film = getById(id);
+            try {
+                if (film != null) {
+                    PreparedStatement pr = conn.prepareStatement("DELETE from t_film WHERE id =" + id + ";");
+                    int updated = pr.executeUpdate();
+                    if (updated > 0) System.out.println("success delete film");
+                    else System.out.println("no film deleted");
+                    pr.close();
+                } else  {
+                    System.out.println("Cannot delete film");
+                }
+            } catch (SQLException e){
+                System.out.println("Failed delete film : " + e.getMessage());
+            }
+    }
+
+    public void getList(ArrayList<Film> list){
+        System.out.println("\nList Film: ");
+        int index = 0;
         try {
-            PreparedStatement pr = conn.prepareStatement("DELETE from t_film WHERE id =" + id + ";");
-            int updated = pr.executeUpdate();
-            if (updated > 0) System.out.println("success delete data");
-            else System.out.println("no data deleted");
-            pr.close();
-
-        } catch (SQLException e){
-            System.out.println("Failed delete : " + e.getMessage());
+            for (Film film: list) {
+                Film films;
+                index++;
+                films = film;
+                System.out.println(films == null ? "There's no film\n" : index +". " + films);
+            }
+        } catch (Exception e) {
+            System.out.println(e.getMessage());
         }
-
     }
 }
